refactor(input): share the remote input-channel vocabulary from shock2vr
The debug runtime's channel patching (head.look, right_hand.trigger, ...)
moves to shock2vr::input::remote so the oculus runtime can speak the same
language on-device. Adds InputOverrides for runtimes that rebuild their
InputContext each frame and must layer patches on top.
