+++ runtimes/debug_runtime/src/main.rs
@@ -32,10 +32,13 @@ use engine::{
 };
 use shock2vr::{
     Game, GameOptions, SpawnLocation,
-    input::{InputAction, InputActionState},
+    input::{InputAction, InputActionState, remote as remote_input},
     input_context::InputContext,
     time::Time,
 };
+// The remote-control channel vocabulary (`head.look`, `right_hand.trigger`, ...)
+// is shared with the oculus runtime's on-device input server.
+use remote_input::{apply_input_patch, head_rotation_from_yaw_pitch};
 
 // Property imports for state queries
 use dark::properties::{PropModelName, PropPosition, PropSymName, PropTemplateId};
@@ -168,26 +171,6 @@ const FIXED_STEP_DT: f32 = 1.0 / 60.0;
 /// so it caps `/v1/step` and `/v1/screenshot` latency at this duration; a few
 /// ms is a no-op for automation but returns the core the rest of the time.
 const IDLE_SLEEP: Duration = Duration::from_millis(4);
-
-/// Head rotation for a yaw/pitch (degrees), matching the desktop runtime's
-/// camera convention (`camera_forward` / `camera_rotation`): at `yaw=pitch=0`
-/// the forward is `(1,0,0)` fed through `look_at_rh`, i.e. looking toward `-X`.
-/// The SAME rotation drives both the render camera and `InputContext.head`, so
-/// the flat viewmodel (placed relative to the head) and the rendered view always
-/// agree - otherwise the viewmodel renders off-axis as a giant side-on slab.
-fn head_rotation_from_yaw_pitch(yaw_deg: f32, pitch_deg: f32) -> Quaternion<f32> {
-    use cgmath::{Decomposed, Rotation, Transform, point3};
-    let (yaw, pitch) = (yaw_deg.to_radians(), pitch_deg.to_radians());
-    let forward = point3(
-        yaw.cos() * pitch.cos(),
-        pitch.sin(),
-        yaw.sin() * pitch.cos(),
-    );
-    let up = vec3(0.0, 1.0, 0.0);
-    let decomposed: Decomposed<Vector3<f32>, Quaternion<f32>> =
-        Transform::look_at_rh(forward, point3(0.0, 0.0, 0.0), up);
-    decomposed.rot.invert()
-}
 
 fn default_camera_head_rotation() -> Quaternion<f32> {
     head_rotation_from_yaw_pitch(0.0, 0.0)
@@ -1813,203 +1796,6 @@ fn input_state_from_context(input: &InputContext) -> commands::InputState {
     }
 }
 
-/// Patch a single channel of the runtime-owned input state, returning false for
-/// an unrecognized channel or a value of the wrong shape. These persist across
-/// frames (a held trigger, a fixed aim), unlike the edge-triggered actions of
-/// `/v1/input/action`.
-///
-/// Recognized channels:
-/// - `head.rotation`                : `[x, y, z, w]` quaternion
-/// - `head.look`                    : `[yaw_deg, pitch_deg]` convenience (forward = -Z).
-///   **Pitch is POSITIVE DOWN** - `+60` looks at the floor, `-60` at the ceiling
-///   - matching the desktop runtime, where mouse-down increments pitch. The same
-///   rotation drives movement, so on a ladder `+pitch` descends and `-pitch`
-///   ascends - a test that pitches `-60` to "look down" at a ladder climbs UP,
-///   which is how issue #603's repro was written.
-/// - `{left,right}_hand.trigger`    : number in [0, 1] (alias `trigger_value`)
-/// - `{left,right}_hand.squeeze`    : number in [0, 1] (alias `squeeze_value`)
-/// - `{left,right}_hand.a`          : number in [0, 1] (alias `a_value`)
-/// - `{left,right}_hand.thumbstick` : `[x, y]`
-/// - `jump`                         : number `0` or `1`
-/// One line describing every recognized input channel, used in error messages so
-/// a bad request is self-documenting. Includes the locomotion semantics (which
-/// stick does what) since that is the game's convention, not guessable.
-fn input_channels_help() -> &'static str {
-    "valid channels: head.rotation [x,y,z,w], head.look [yaw_deg,pitch_deg], \
-     pointer.position [x,y] in [0,1] (origin top-left; null clears), pointer.pressed 0|1, \
-     {left,right}_hand.{trigger,squeeze,a} <number 0..1>, \
-     {left,right}_hand.thumbstick [x,y], \
-     {left,right}_hand.position [x,y,z] (pawn-local), \
-     {left,right}_hand.rotation [x,y,z,w], \
-     crouch 0|1 (stand-up refused without headroom), \
-     jump 0|1 (held; launches on the grounded rising edge); \
-     locomotion: right_hand.thumbstick [strafe, forward] moves the player, \
-     left_hand.thumbstick.x turns, left_hand.thumbstick.y flies up/down"
-}
-
-fn apply_input_patch(input: &mut InputContext, channel: &str, value: &Value) -> Result<(), String> {
-    use cgmath::Vector2;
-
-    // Parse a scalar/array value for `channel`, attributing a clear error to the
-    // channel and value shape when it doesn't match.
-    fn num(channel: &str, v: &Value) -> Result<f32, String> {
-        v.as_f64()
-            .map(|f| f as f32)
-            .ok_or_else(|| format!("channel '{channel}' expects a number, got {v}"))
-    }
-    fn arr(channel: &str, v: &Value, n: usize) -> Result<Vec<f32>, String> {
-        let a = v.as_array().ok_or_else(|| {
-            format!("channel '{channel}' expects an array of {n} numbers, got {v}")
-        })?;
-        if a.len() != n {
-            return Err(format!(
-                "channel '{channel}' expects {n} numbers, got {} ({v})",
-                a.len()
-            ));
-        }
-        a.iter()
-            .map(|e| num(channel, e))
-            .collect::<Result<Vec<_>, _>>()
-    }
-
-    // Hand channels: "<left|right>_hand.<field>"
-    if let Some((side, field)) = channel.split_once("_hand.") {
-        let hand = match side {
-            "left" => &mut input.left_hand,
-            "right" => &mut input.right_hand,
-            _ => {
-                return Err(format!(
-                    "unknown input channel '{channel}'; {}",
-                    input_channels_help()
-                ));
-            }
-        };
-        return match field {
-            "trigger" | "trigger_value" => {
-                hand.trigger_value = num(channel, value)?;
-                Ok(())
-            }
-            "squeeze" | "squeeze_value" => {
-                hand.squeeze_value = num(channel, value)?;
-                Ok(())
-            }
-            "a" | "a_value" => {
-                hand.a_value = num(channel, value)?;
-                Ok(())
-            }
-            "thumbstick" => {
-                let a = arr(channel, value, 2)?;
-                hand.thumbstick = Vector2::new(a[0], a[1]);
-                Ok(())
-            }
-            "position" => {
-                let p = arr(channel, value, 3)?;
-                hand.position = cgmath::vec3(p[0], p[1], p[2]);
-                Ok(())
-            }
-            "rotation" => {
-                let q = arr(channel, value, 4)?;
-                hand.rotation =
-                    cgmath::InnerSpace::normalize(Quaternion::new(q[3], q[0], q[1], q[2]));
-                Ok(())
-            }
-            _ => Err(format!(
-                "unknown input channel '{channel}'; {}",
-                input_channels_help()
-            )),
-        };
-    }
-
-    match channel {
-        "head.rotation" => {
-            let q = arr(channel, value, 4)?;
-            input.head.rotation = Quaternion::new(q[3], q[0], q[1], q[2]);
-            Ok(())
-        }
-        // Desktop camera convention (yaw=pitch=0 looks toward -X); drives both
-        // the render camera and the viewmodel. Convenient for pointing the
-        // camera without hand-authoring a quaternion.
-        "head.look" => {
-            let yp = arr(channel, value, 2)?;
-            input.head.rotation = head_rotation_from_yaw_pitch(yp[0], yp[1]);
-            Ok(())
-        }
-        // Flat-mode 2D pointer (cursor). Position is normalized [0,1] per
-        // axis, origin top-left; setting either channel materializes the
-        // pointer (it is `None` until first set). `pointer.position: null`
-        // clears the pointer back to `None` (scenes branch on Some/None, so
-        // the cleared state must be reachable for testing).
-        "pointer.position" => {
-            if value.is_null() {
-                input.pointer = None;
-                return Ok(());
-            }
-            let xy = arr(channel, value, 2)?;
-            if !(xy[0].is_finite() && xy[1].is_finite())
-                || !(0.0..=1.0).contains(&xy[0])
-                || !(0.0..=1.0).contains(&xy[1])
-            {
-                return Err(format!(
-                    "channel '{channel}' expects normalized coordinates in [0,1], got [{}, {}]",
-                    xy[0], xy[1]
-                ));
-            }
-            let pointer = input
-                .pointer
-                .get_or_insert(shock2vr::input_context::Pointer2D {
-                    position: cgmath::vec2(0.0, 0.0),
-                    pressed: false,
-                });
-            pointer.position = cgmath::vec2(xy[0], xy[1]);
-            Ok(())
-        }
-        "pointer.pressed" => {
-            let pressed = match num(channel, value)? {
-                v if v == 0.0 => false,
-                v if v == 1.0 => true,
-                v => {
-                    return Err(format!("channel '{channel}' expects 0 or 1, got {v}"));
-                }
-            };
-            let pointer = input
-                .pointer
-                .get_or_insert(shock2vr::input_context::Pointer2D {
-                    position: cgmath::vec2(0.0, 0.0),
-                    pressed: false,
-                });
-            pointer.pressed = pressed;
-            Ok(())
-        }
-        // Crouch request: like the desktop LeftControl hold. The ACTUAL state
-        // can lag (standing up is refused without headroom); observe it via
-        // the player's body y (the collider center drops when crouched).
-        "crouch" => {
-            input.crouch = match num(channel, value)? {
-                v if v == 0.0 => false,
-                v if v == 1.0 => true,
-                v => {
-                    return Err(format!("channel '{channel}' expects 0 or 1, got {v}"));
-                }
-            };
-            Ok(())
-        }
-        "jump" => {
-            input.jump = match num(channel, value)? {
-                v if v == 0.0 => false,
-                v if v == 1.0 => true,
-                v => {
-                    return Err(format!("channel '{channel}' expects 0 or 1, got {v}"));
-                }
-            };
-            Ok(())
-        }
-        _ => Err(format!(
-            "unknown input channel '{channel}'; {}",
-            input_channels_help()
-        )),
-    }
-}
-
 fn input_snapshot_from_context(input: &InputContext) -> InputSnapshot {
     fn hand_snapshot(hand: commands::InputHand) -> HandSnapshot {
         HandSnapshot {
@@ -3469,49 +3255,13 @@ async fn get_input_state(
     }
 }
 
-/// Parse a `/v1/control/input` POST body into input patches, accepting either
-/// shape:
-/// - explicit:  `{"channel": "right_hand.trigger", "value": 1.0}`
-/// - map form:  `{"right_hand.trigger": 1.0, "head.look": [30, 0]}`
-///
-/// The map form is detected when the object does NOT have both `channel` and
-/// `value` keys. Returns an actionable error (not a silent empty patch) when the
-/// body isn't a usable object, so a malformed request fails loudly.
+/// Parse a `/v1/control/input` POST body into input patches. The channel
+/// vocabulary and both accepted body shapes live in `shock2vr::input::remote`,
+/// shared with the oculus runtime's remote-input server.
 fn parse_input_patches(body: &serde_json::Value) -> Result<Vec<commands::InputPatch>, String> {
-    let obj = body.as_object().ok_or_else(|| {
-        format!(
-            "request body must be a JSON object, e.g. {{\"channel\":\"right_hand.trigger\",\"value\":1.0}} \
-             or {{\"right_hand.trigger\":1.0}}; {}",
-            input_channels_help()
-        )
-    })?;
-
-    // Explicit {channel, value} form.
-    if obj.contains_key("channel") && obj.contains_key("value") {
-        let channel = obj["channel"]
-            .as_str()
-            .ok_or_else(|| "\"channel\" must be a string".to_string())?
-            .to_string();
-        return Ok(vec![commands::InputPatch {
-            channel,
-            value: obj["value"].clone(),
-        }]);
-    }
-
-    if obj.is_empty() {
-        return Err(format!(
-            "no input channels in request; {}",
-            input_channels_help()
-        ));
-    }
-
-    // Map form: every key is a channel name.
-    Ok(obj
-        .iter()
-        .map(|(channel, value)| commands::InputPatch {
-            channel: channel.clone(),
-            value: value.clone(),
-        })
+    Ok(remote_input::parse_input_patches(body)?
+        .into_iter()
+        .map(|(channel, value)| commands::InputPatch { channel, value })
         .collect())
 }
