+++ shock2vr/src/input/mod.rs
@@ -1,5 +1,6 @@
 mod actions;
 mod dispatcher;
+pub mod remote;
 mod state;
 
 pub use actions::*;
